module_init: Replace module_init with late_initcall
When we install p2pmem-pci.ko into a kernel there is a chance it will
load before the device driver(s) for the PCIe devices that are
publishing p2pdma memory (nvme for example). Therefore we want this
module to load as late as possible via a late_initcall().

Fixes #2.